content api improvements
